Reland: Add flag pace graph and fix misc. issues (#19)
This change primarily adds a "flag pace" page that shows a graph of scores over round time. It uses d3js running in the Node environment to statically create the SVG markup and pass it to the template.

In order to use d3js from within Node, I had to move the to ES Modules in Node instead of CommonJS. A variety of small changes were made to support that.

Other misc. fixes/improvements:
* Fix flag time handling for when both teams have flag movement (e.g. as would happen in an 8v8)
* Add conc kills column to O/D comparisons on summary page
* Fix O/D comparison CSS styling